requirements.txt
boto3==1.13.15
pyyaml==5.3.1